requirements.txt
Django==1.6.11
MySQL-python==1.2.5
Pillow==2.8.1
South==1.0.2
Unipath==1.1
ipython==3.1.0
sorl-thumbnail==12.2
